+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## develop
+### Added
+- Demo mode: "demo" badge and instructions added
+- Small fix in title
+- Fix behaviour for register form
+
 ## [0.5.0] - 2022-01-04
 ### Added
 - Demonstration mode with dummy data